Fix issues with division
The previous division impl was pretty sketchy and the random numbers
generated for benchmarking were hitting some cases that weren't working
anyway.

Implements 3-by-2 limb division to complement the existing 2-by-1
division, and a few other required helper routines. The new "Schoolbook"
division algorithm relies on the increased accuracy of 3-by-2 limb
division to avoid doing as much adjustment of the quotient.

An extra test, `div_rand`, has been added that tests that the results
are correct on randomly-generated integers.