Implement `pow` more efficiently in the `ll` module.
This moves the implementation of `pow` on `int` to the `ll` module. This
has resulted in a reasonable performance increas, especially for
smaller numbers. The "m-ary" branch was not ported as it was not
actually correct.

Also removes the crate-level `#![allow(dead_code)]` and adds some
function specific ones were appropriate.