Fix conversion to and from minimum-value signed integers
Normally we can just take the absolute value of a signed integer when
doing the conversion, however that isn't possible when the integer is
the minimum value. As such, it needs to be handled specially.

Converting to such a value also requires care, as `int::MIN * -1`
overflows, fortunately using wrapping muliplication accounts for this
nicely.

Fixes #20