Add a specific sqr function for squaring integers
Squaring can be done slightly more efficiently than regular
multiplication, so there is now a special case for it used by the
square/dsquare methods on Int and in the pow implementation.

Right now it is not much faster than regular multiplication, however the
potential for improvements as more sophisticated techniques are
implemented should mean that users of sqr will automatically see an
improvement.