Reimplement BitOr and BitAnd.
Share code between the two operations as much as possible.

It seems the best way to implement bitwise operations on negative
big-ints is to literally convert to two's complement and then convert
back if needed, at the end. Unfortunately that means mutating both
arguments (in general).

This implementation tries to maintain speed by taking a fast path when
both arguments are positive.

Fixes #26.