Optimise `ll::base::num_base_digits` for 2 & other powers of it.
The old code would end up with an actual division when inlined into
another crate, because the `BASES` symbol isn't inlined and hence the
compiler can't tell that the division will be by a power of two. This
caused `bit_length` to be very high in the profiles of my `float` crate
with the `div` instruction taking the majority of the time. It's
especially unfortunate because the divison for `bit_length` is a
division by 1.

(A simple benchmark of `Int::bit_length` shows the performance dropping
from 9 ns/iter to 1-2 ns/iter, with `div` taking ~80% of the time in the
former case.)

cc #52