Fix calculation of bit length in `gen_uint_below`
`gen_uint_below` was incorrectly calculating the number of bits required
to represent `bound`, causing it to calculate too few bits in some cases
and too many in others.

In the case where is would produce a too-large value, this could cause
the number of invalid values produced by `gen_uint` to rise
dramatically, effectively hanging the program.

The code now uses the `bit_length` method to calculate the number of
bits.

Fixes #58